+++ src/api/rest/gwflow.py
@@ -98,6 +98,35 @@ def zonesId(id):
 
 	abort(405, 'HTTP Method not allowed.')
 
+# Gwflow_cell
+
+@bp.route('/cells', methods=['GET'])
+def cells():
+	if request.method == 'GET':
+		table = gwflow.Gwflow_cell
+		filter_cols = [table.cell_id]
+		return DefaultRestMethods.get_paged_list(table, filter_cols)
+	abort(405, 'HTTP Method not allowed.')
+
+@bp.route('/cells/<int:id>', methods=['GET', 'PUT'])
+def cellsId(id):
+	project_db = request.headers.get(rh.PROJECT_DB)
+	has_db,error = rh.init(project_db)
+	if not has_db: abort(400, error)
+	table = gwflow.Gwflow_cell
+
+	if request.method == 'GET':
+		m = table.get_or_none(table.cell_id == id)
+		if m is None:
+			rh.close()
+			abort(400, '{} {} does not exist'.format('Cell', id))
+		rh.close()
+		return model_to_dict(m, recurse=False)
+	elif request.method == 'PUT':
+		return DefaultRestMethods.put(id, gwflow.Gwflow_cell, 'Cell', primary_key='cell_id')
+
+	abort(405, 'HTTP Method not allowed.')
+
 
 # Gwflow_fpcell
 
@@ -115,17 +144,27 @@ def fpcell():
 		return delete_all_cells(table, description)
 	abort(405, 'HTTP Method not allowed.')
 
-@bp.route('/fpcell/<int:id>', methods=['GET', 'PUT', 'DELETE'])
-def fpcellId(id):
+@bp.route('/fpcell/<int:id>', methods=['PUT'])
+def fpcellIdPut(id):
 	table = gwflow.Gwflow_fpcell
 	description = 'Floodplain cell'
 
+	if request.method == 'PUT':
+		return put_cell(id, table, description) # Note, they are allowed to change the channel in the form, so don't include the original channel id in this method
+
+	abort(405, 'HTTP Method not allowed.')
+
+@bp.route('/fpcell/<int:id>/<int:channel_id>', methods=['GET', 'DELETE'])
+def fpcellId(id, channel_id):
+	table = gwflow.Gwflow_fpcell
+	composite_key_part2_field = gwflow.Gwflow_fpcell.channel_id
+	composite_key_part2_value = channel_id
+	description = 'Floodplain cell'
+
 	if request.method == 'GET':
-		return get_cell(table, id, description)
+		return get_cell(table, id, description, composite_key_part2_field, composite_key_part2_value)
 	elif request.method == 'DELETE':
-		return delete_cell(table, id, description)
-	elif request.method == 'PUT':
-		return put_cell(id, table, description)
+		return delete_cell(table, id, description, composite_key_part2_field, composite_key_part2_value)
 
 	abort(405, 'HTTP Method not allowed.')
 
@@ -145,17 +184,27 @@ def rescell():
 		return delete_all_cells(table, description)
 	abort(405, 'HTTP Method not allowed.')
 
-@bp.route('/rescell/<int:id>', methods=['GET', 'PUT', 'DELETE'])
-def rescellId(id):
+@bp.route('/rescell/<int:id>', methods=['PUT'])
+def rescellIdPut(id):
 	table = gwflow.Gwflow_rescell
 	description = 'Reservoir cell'
 
+	if request.method == 'PUT':
+		return put_cell(id, table, description) # Note, they are allowed to change the reservoir in the form, so don't include the original reservoir id in this method
+
+	abort(405, 'HTTP Method not allowed.')
+
+@bp.route('/rescell/<int:id>/<int:reservoir_id>', methods=['GET', 'DELETE'])
+def rescellId(id, reservoir_id):
+	table = gwflow.Gwflow_rescell
+	composite_key_part2_field = gwflow.Gwflow_rescell.reservoir_id
+	composite_key_part2_value = reservoir_id
+	description = 'Reservoir cell'
+
 	if request.method == 'GET':
-		return get_cell(table, id, description)
+		return get_cell(table, id, description, composite_key_part2_field, composite_key_part2_value)
 	elif request.method == 'DELETE':
-		return delete_cell(table, id, description)
-	elif request.method == 'PUT':
-		return put_cell(id, table, description)
+		return delete_cell(table, id, description, composite_key_part2_field, composite_key_part2_value)
 
 	abort(405, 'HTTP Method not allowed.')
 
@@ -329,8 +378,8 @@ def solutes():
 		return DefaultRestMethods.get_paged_list(table, filter_cols)
 	abort(405, 'HTTP Method not allowed.')
 
-@bp.route('/solutes/<name>', methods=['GET', 'PUT'])
-def solutesId(name):
+@bp.route('/solutes/<int:id>', methods=['GET', 'PUT'])
+def solutesId(id):
 	table = gwflow.Gwflow_solute
 	description = 'Solute'
 
@@ -339,10 +388,10 @@ def solutesId(name):
 	if not has_db: abort(400, error)
 
 	if request.method == 'GET':
-		m = table.get_or_none(table.name == name)
+		m = table.get_or_none(table.id == id)
 		if m is None:
 			rh.close()
-			abort(400, '{} {} does not exist'.format(description, name))
+			abort(400, '{} {} does not exist'.format(description, id))
 		rh.close()
 		return model_to_dict(m, recurse=False)
 	elif request.method == 'PUT':
@@ -353,14 +402,14 @@ def solutesId(name):
 				rate_const=args.get('rate_const', 0),
 				canal_irr=args.get('canal_irr', 0),
 				init_conc=args.get('init_conc', 0),
-			).where(table.name == name).execute()
+			).where(table.id == id).execute()
 			rh.close()
 			if result > 0:
 				return '', 200
-			abort(400, 'Unable to update {}.'.format(name))
+			abort(400, 'Unable to update {}.'.format(id))
 		except table.DoesNotExist:
 			rh.close()
-			abort(400, '{} {} does not exist'.format(description, name))
+			abort(400, '{} {} does not exist'.format(description, id))
 		except Exception as ex:
 			rh.close()
 			abort(400, 'Unexpected error {ex}'.format(ex=ex))
@@ -405,10 +454,10 @@ def solutesDefault():
 
 gis_cols = {
 	gwflow.Gwflow_hrucell: ['hru_id', connect.Hru_con, 'many', False],
-	gwflow.Gwflow_fpcell: ['channel_id', connect.Chandeg_con, 'single', True],
+	gwflow.Gwflow_fpcell: ['channel_id', connect.Chandeg_con, 'many', True],
 	gwflow.Gwflow_chancell: ['channel_id', connect.Chandeg_con, 'single', False],
 	gwflow.Gwflow_lsucell: ['lsu_id', connect.Rout_unit_con, 'many', False],
-	gwflow.Gwflow_rescell: ['reservoir_id', connect.Reservoir_con, 'single', True],
+	gwflow.Gwflow_rescell: ['reservoir_id', connect.Reservoir_con, 'many', True],
 }
 
 def get_cell_paged(table, filter_cols) -> Response:
@@ -418,13 +467,15 @@ def get_cell_paged(table, filter_cols) -> Response:
 	gis_col = gis_cols.get(table, None)
 
 	ml = [model_to_dict(v, recurse=False) for v in m]
-	gis_to_con_name = IndexHelper(gis_col[1]).get_names()
-	for d in ml:
-		id = d[gis_col[0]]
-		d[gis_col[0]] = {
-			'id': id,
-			'name': gis_to_con_name.get(id, None)
-		}
+
+	if gis_col is not None:
+		gis_to_con_name = IndexHelper(gis_col[1]).get_names()
+		for d in ml:
+			id = d[gis_col[0]]
+			d[gis_col[0]] = {
+				'id': id,
+				'name': gis_to_con_name.get(id, None)
+			}
 
 	return {
 		'total': items['total'],
@@ -432,25 +483,27 @@ def get_cell_paged(table, filter_cols) -> Response:
 		'items': ml
 	}
 
-def get_cell(table, id, description) -> Response:
+def get_cell_model(table, id, composite_key_part2_field = None, composite_key_part2_value = None):
+	if composite_key_part2_field is not None and composite_key_part2_value is not None:
+		return table.get_or_none((table.cell_id == id) & (composite_key_part2_field == composite_key_part2_value))
+	else:
+		return table.get_or_none(table.cell_id == id)
+
+def get_cell(table, id, description, composite_key_part2_field = None, composite_key_part2_value = None) -> Response:
 	project_db = request.headers.get(rh.PROJECT_DB)
 	has_db,error = rh.init(project_db)
 	if not has_db: abort(400, error)
 
-	m = table.get_or_none(table.cell_id == id)
+	m = get_cell_model(table, id, composite_key_part2_field, composite_key_part2_value)
 	if m is None:
 		rh.close()
 		abort(400, '{} {} does not exist'.format(description, id))
 
 	gis_col = gis_cols.get(table, None)
 	d = model_to_dict(m, recurse=False)
-	gis_to_con_name = IndexHelper(gis_col[1]).get_names()
-	gid = d[gis_col[0]]
-	d[gis_col[0]] = {
-		'id': gid,
-		'name': gis_to_con_name.get(gid, None)
-	}
-
+	con = gis_col[1].get_or_none(gis_col[1].gis_id == d[gis_col[0]])
+	d['{}_name'.format(gis_col[0])] = con.name if con is not None else None
+	d.pop(gis_col[0], None)
 	rh.close()
 	return d
 
@@ -463,17 +516,8 @@ def post_cell(table, description) -> Response:
 	gis_col = gis_cols.get(table, None)
 
 	try:
-		m = table()
-		m.cell_id = args['cell_id']
-		if gis_col is not None and gis_col[0] in args:
-			setattr(m, gis_col[0], args[gis_col[0]])
+		result = save_cell_args(table, request.json, is_new=True)
 
-		for key in args:
-			if key not in ('cell_id', gis_col[0]) if gis_col else ('cell_id',):
-				if hasattr(m, key):
-					setattr(m, key, args[key])
-
-		result = m.save()
 		rh.close()
 		if result > 0:
 			return {'id': result}, 201
@@ -489,7 +533,8 @@ def put_cell(id, table, description) -> Response:
 	if not has_db: abort(400, error)
 
 	try:
-		result = RestHelpers.save_args(table, request.json, id=id)
+		result = save_cell_args(table, request.json, id=id)
+
 		rh.close()
 		if result > 0:
 			return '', 200
@@ -501,12 +546,15 @@ def put_cell(id, table, description) -> Response:
 		rh.close()
 		abort(400, 'Unexpected error {ex}'.format(ex=ex))
 
-def delete_cell(table, id, description) -> Response:
+def delete_cell(table, id, description, composite_key_part2_field = None, composite_key_part2_value = None) -> Response:
 	project_db = request.headers.get(rh.PROJECT_DB)
 	has_db,error = rh.init(project_db)
 	if not has_db: abort(400, error)
 
-	result = table.delete().where(table.cell_id == id).execute()
+	if composite_key_part2_field is not None and composite_key_part2_value is not None:
+		result = table.delete().where((table.cell_id == id) & (composite_key_part2_field == composite_key_part2_value)).execute()
+	else:
+		result = table.delete().where(table.cell_id == id).execute()
 	rh.close()
 	if result > 0:
 		return '', 204
@@ -522,3 +570,27 @@ def delete_all_cells(table, description) -> Response:
 	if result > 0:
 		return '', 204
 	abort(400, 'Error deleting {}.'.format(description.lower()))
+
+def save_cell_args(table, args, id=0, is_new=False):
+	gis_col = gis_cols.get(table, None)
+	gis_name_to_con = IndexHelper(gis_col[1]).get_id_from_name()
+	gis_name = '{}_name'.format(gis_col[0]) # Note, this is simply a UI decision what this is called, so in RescellForm.vue for example, it's just reservoir_id_name for lookup simplicity.
+
+	params = {}
+	table_gis_field = None
+	for field in table._meta.sorted_fields:
+		if field.name in args or '{}_name'.format(field.name) in args:
+			if field.name == gis_col[0]:
+				table_gis_field = field
+				params[field.name] = gis_name_to_con.get(args[gis_name], None)
+			else:
+				params[field.name] = args[field.name]
+
+	if is_new:
+		query = table.insert(params)
+	elif gis_col[2] == 'single':
+		query = table.update(params).where(table.cell_id == id)
+	else:
+		query = table.update(params).where((table.cell_id == id) & (table_gis_field == params[gis_col[0]]))
+	
+	return query.execute()