+++ src/renderer/views/edit/connect/gwflow/FpcellEdit.vue
@@ -16,12 +16,12 @@
 	});
 
 	async function get() {
-		if (route.params.id === undefined) return;
+		if (route.params.id === undefined || route.params.comp_id === undefined) return;
 		data.page.loading = true;
 		data.page.error = null;
 
 		try {
-			const response = await api.get(`gwflow/fpcell/${route.params.id}`, currentProject.getApiHeader());
+			const response = await api.get(`gwflow/fpcell/${route.params.id}/${route.params.comp_id}`, currentProject.getApiHeader());
 			errors.log(response.data);
 
 			data.item = response.data;
@@ -35,11 +35,12 @@
 	onMounted(async () => await get())
 	watch(() => route.name, async () => await get())
 	watch(() => route.params.id, async () => await get())
+	watch(() => route.params.comp_id, async () => await get())
 </script>
 
 <template>
 	<project-container :loading="data.page.loading" :load-error="data.page.error">
-		<file-header input-file="gwflow.floodplain" docs-path="modflow" use-io>
+		<file-header input-file="floodplain.gw" docs-path="modflow" use-io>
 			<router-link to="/edit/cons/gwflow">Groundwater Flow</router-link>
 			/ <router-link to="/edit/cons/gwflow/fpcell">Floodplain</router-link>
 			/ Edit