+++ changelog.md
@@ -2,6 +2,14 @@
 
 ## Version 3 ##
 
+### Revision 4.0.0 ###
+**IMPORTANT:** This new version of the editor is only compatible with SWAT+ rev. 62 and later. Due to structural model changes, rev. 61 and earlier are NOT supported. Project updates are available after software update.
+
+* Compatible with SWAT+ rev. 62
+* gwflow structure updates (QSWAT+ v4.0 update is REQUIRED)
+* Add carbon module (see basin section for carbon and carbon layers)
+* Update print.prt to include gwflow options and legacy carbon options
+
 ### Revision 3.2.4 ###
 
 * Add fix for resetting observed weather input directory when stations are removed